refactor: 関数の命名に統一性を持たせるためにPrefLocationを追記
